[ReactCore] Share WalletConnecthandler in react core
