+++ packages/react-core/src/core/utils/receiver.ts
@@ -1,0 +1,10 @@
+import { WalletInstance } from "../types/wallet";
+
+export const isWalletConnectReceiverEnabled = (wallet?: WalletInstance) => {
+  const options = wallet?.getOptions();
+  return (
+    options &&
+    "walletConnectReceiver" in options &&
+    options.walletConnectReceiver
+  );
+};
