+++ packages/react-core/src/evm/index.ts
@@ -23,6 +23,7 @@ export {
 export { useNetwork } from "../core/hooks/useNetwork";
 export { useWalletContext } from "../core/providers/thirdweb-wallet-provider";
 export { usePersonalWalletAddress } from "./hooks/usePersonalWalletAddress";
+export { useWalletConnectHandler } from "../core/providers/thirdweb-wallet-provider";
 
 // connected wallet hooks
 export {
